Move passwordless Identity server files to Admin
