+++ src/Admin/Controllers/LoginController.cs
@@ -1,5 +1,5 @@
-﻿using Bit.Admin.Models;
-using Bit.Core.Identity;
+﻿using Bit.Admin.IdentityServer;
+using Bit.Admin.Models;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
