+++ src/Admin/IdentityServer/ReadOnlyEnvIdentityUserStore.cs
@@ -1,8 +1,7 @@
 ﻿using Bit.Core.Utilities;
 using Microsoft.AspNetCore.Identity;
-using Microsoft.Extensions.Configuration;
 
-namespace Bit.Core.Identity;
+namespace Bit.Admin.IdentityServer;
 
 public class ReadOnlyEnvIdentityUserStore : ReadOnlyIdentityUserStore
 {
@@ -14,7 +13,7 @@ public class ReadOnlyEnvIdentityUserStore : ReadOnlyIdentityUserStore
     }
 
     public override Task<IdentityUser> FindByEmailAsync(string normalizedEmail,
-        CancellationToken cancellationToken = default(CancellationToken))
+        CancellationToken cancellationToken = default)
     {
         var usersCsv = _configuration["adminSettings:admins"];
         if (!CoreHelpers.SettingHasValue(usersCsv))
@@ -59,7 +58,7 @@ public class ReadOnlyEnvIdentityUserStore : ReadOnlyIdentityUserStore
     }
 
     public override Task<IdentityUser> FindByIdAsync(string userId,
-        CancellationToken cancellationToken = default(CancellationToken))
+        CancellationToken cancellationToken = default)
     {
         return FindByEmailAsync(userId, cancellationToken);
     }
