+++ src/SharedWeb/Utilities/ServiceCollectionExtensions.cs
@@ -33,7 +33,6 @@ using Microsoft.Extensions.Caching.Distributed;
 using Microsoft.Extensions.Caching.StackExchangeRedis;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Options;
 using Serilog.Context;
@@ -354,39 +353,7 @@ public static class ServiceCollectionExtensions
         return identityBuilder;
     }
 
-    public static Tuple<IdentityBuilder, IdentityBuilder> AddPasswordlessIdentityServices<TUserStore>(
-        this IServiceCollection services, GlobalSettings globalSettings) where TUserStore : class
-    {
-        services.TryAddTransient<ILookupNormalizer, LowerInvariantLookupNormalizer>();
-        services.Configure<DataProtectionTokenProviderOptions>(options =>
-        {
-            options.TokenLifespan = TimeSpan.FromMinutes(15);
-        });
 
-        var passwordlessIdentityBuilder = services.AddIdentity<IdentityUser, Role>()
-            .AddUserStore<TUserStore>()
-            .AddRoleStore<RoleStore>()
-            .AddDefaultTokenProviders();
-
-        var regularIdentityBuilder = services.AddIdentityCore<User>()
-            .AddUserStore<UserStore>();
-
-        services.TryAddScoped<PasswordlessSignInManager<IdentityUser>, PasswordlessSignInManager<IdentityUser>>();
-
-        services.ConfigureApplicationCookie(options =>
-        {
-            options.LoginPath = "/login";
-            options.LogoutPath = "/";
-            options.AccessDeniedPath = "/login?accessDenied=true";
-            options.Cookie.Name = $"Bitwarden_{globalSettings.ProjectName}";
-            options.Cookie.HttpOnly = true;
-            options.ExpireTimeSpan = TimeSpan.FromDays(2);
-            options.ReturnUrlParameter = "returnUrl";
-            options.SlidingExpiration = true;
-        });
-
-        return new Tuple<IdentityBuilder, IdentityBuilder>(passwordlessIdentityBuilder, regularIdentityBuilder);
-    }
 
     public static void AddIdentityAuthenticationServices(
         this IServiceCollection services, GlobalSettings globalSettings, IWebHostEnvironment environment,
